Chapter 4 Finished with all controllers done

+++ shop/src/main/java/ro/msg/learning/shop/service/strategies/SingleLocationStrategy.java
@@ -36,7 +36,7 @@ public class SingleLocationStrategy implements IWhichStrategy {
                         .findFirst()
                         .orElse(null);
 
-                if (newStock != null && newStock.getQuantity() > (Integer) stock.getValue() &&
+                if (newStock != null && newStock.getQuantity() > stock.getValue() &&
                         !productsStocks.contains(newStock)) {
                     productsStocks.add(newStock);
                 }
